Add treatment for inexisting id on Post, Put and Delete

+++ forum/src/main/java/br/com/andrezasecon/forum/controllers/TopicosController.java
@@ -17,6 +17,7 @@ import javax.validation.Valid;
 import java.io.Serializable;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController // Adiciona automaticamente o @ResponseBody em cada método
 @RequestMapping(value = "/topicos")
@@ -54,23 +55,35 @@ public class TopicosController implements Serializable {
     }
 
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDto detail(@PathVariable Long id){
-        Topico topico = topicoRepository.getById(id);
-        return new DetalhesDoTopicoDto(topico);
+    public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id){
+       Optional<Topico> topico = topicoRepository.findById(id);
+       if(topico.isPresent()) {
+           return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+       }
+       return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     @Transactional // faz o JPA commitar a transação no banco de dados
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-       Topico topico = form.atualizar(id, topicoRepository);
-       return ResponseEntity.ok(new TopicoDto(topico));
+       Optional<Topico> optional = topicoRepository.findById(id);
+       if(optional.isPresent()) {
+           Topico topico = form.atualizar(id, topicoRepository);
+           return ResponseEntity.ok(new TopicoDto(topico));
+        }
+        return ResponseEntity.notFound().build();
+
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id){
-        topicoRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if(optional.isPresent()) {
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 
 }
